requirements.txt
langchain
langchain-ollama
python-dotenv
langchain-text-splitters
pypdf
langchain-core
langchain-postgres
fastapi
uvicorn
python-multipart
alembic
sqlalchemy
asyncpg
psycopg2-binarypython-jose[cryptography]
passlib[bcrypt]
